Muslim Prayer App
This application let's the user choose a city/location to retrieve the muslim prayer times of the day.

The preferred location will stay the same until the user changes it.

Using Calendar View the user can choose for which date they want to retrieve the prayer times for (for the selected location).

The application will show all prayer times, the sunrise time and also calculate & show time to next prayer and calculate & show the time for which half of the night has passed.

The Isha prayer has to be performed before half of the night has passed.